Add support for Leapfrog Creatr HS with example config files
The Configuration*.h files in the example directory will configure
Marlin for the production Creatr HS printer (with corrected extrusion
and X/Y steps).

No changes were made to Configuration_adv.h, but I copied it to the
example directory for completeness in case something changes later on.

Not sure what board the LEAPFROG is, the only difference I found was
the fan being on a different pin and it has a (bogus?) heater2 defined.
The Creatr HS board is internally jumpered to drive both Y steppers
from one output so Y2 isn’t used, the same pins are used for second
extruder.